Fix float not being stored as uint32_t correctly

+++ inc/libutil/remote_var_manager.h
@@ -66,7 +66,7 @@ public:
 
 		void SetReal(const float val)
 		{
-			m_val = *reinterpret_cast<const float*>(&val);
+			m_val = *reinterpret_cast<const uint32_t*>(&val);
 		}
 
 	private:
